fix: Close keyboard without confirm button
Enhances the user experience in the add and edit record dialogs
by dismissing the keyboard when tapping outside the input fields.

Also, removes redundant calls to clear snackbars, as they are
implicitly handled by showing a new snackbar.

Fixes an issue in the edit record dialog where the `onAdd` function
was not correctly referencing `this`, and binds the function to
the current instance.